transplant bfd/daemon and bfd/cli into mgd and mgadm
