+++ Cargo.toml
@@ -6,10 +6,7 @@ default-members = [
     "ddmadm",
     "ddmd",
     "ddm-admin-client",
-    "bfd/lib",
-    "bfd/daemon",
-    "bfd/client",
-    "bfd/cli",
+    "bfd",
     "util",
     "mg-ddm-verify",
     "tests",
@@ -28,10 +25,7 @@ members = [
     "ddmadm",
     "ddmd",
     "ddm-admin-client",
-    "bfd/lib",
-    "bfd/daemon",
-    "bfd/client",
-    "bfd/cli",
+    "bfd",
     "package",
     "util",
     "mg-ddm-verify",
