+++ bfd/Cargo.toml
@@ -7,7 +7,6 @@ edition = "2021"
 slog.workspace = true
 slog-bunyan.workspace = true
 slog-async.workspace = true
-thiserror.workspace = true
 num_enum.workspace = true
 anyhow.workspace = true
 schemars.workspace = true
