--- bfd/client/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "bfd-client"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-progenitor.workspace = true
-reqwest.workspace = true
-serde.workspace = true
-slog.workspace = true
