--- bfd/daemon/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "bfdd"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-bfd = { path = "../lib" }
-clap.workspace = true
-tokio.workspace = true
-slog.workspace = true
-slog-bunyan.workspace = true
-slog-async.workspace = true
-dropshot.workspace = true
-anyhow.workspace = true
-anstyle.workspace = true
-serde.workspace = true
-schemars.workspace = true
-socket2.workspace = true
