+++ bfd/src/util.rs
