+++ mgd/Cargo.toml
@@ -6,8 +6,10 @@ edition = "2021"
 [dependencies]
 mg-lower = { path = "../mg-lower", optional = true }
 mg-common = { path = "../mg-common" }
+bfd = { path = "../bfd" }
 bgp = { path = "../bgp" }
 rdb = { path = "../rdb" }
+anyhow.workspace = true
 clap.workspace = true
 colored.workspace = true
 dropshot.workspace = true
@@ -18,6 +20,7 @@ slog-async.workspace = true
 slog-term.workspace = true
 tokio.workspace = true
 http.workspace = true
+socket2.workspace = true
 thiserror.workspace = true
 rand.workspace = true
